+++ packages/atomic/src/components/commerce/search-box-suggestions/atomic-commerce-search-box-recent-queries/atomic-commerce-search-box-recent-queries.spec.ts
@@ -1,0 +1,20 @@
+import {describe, test} from 'vitest';
+
+//TODO: Write those tests during the lit migration
+//https://coveord.atlassian.net/browse/KIT-3958
+describe('AtomicCommerceSearchBoxRecentQueries', () => {
+  describe('when clicking', () => {
+    test('on a recent query, it should execute the proper search', async () => {});
+    test('on the "Clear" button, it should clear the recent queries', async () => {});
+  });
+
+  describe('when using keyboard navigation', () => {
+    test('on recent queries, it should reflect the proper query in the search box', async () => {});
+    test('on recent queries, it should execute the proper search', async () => {});
+    test('on the "Clear" button, it should clear the recent queries ', async () => {});
+  });
+
+  describe('when using keyboard navigation and then switching to mouse and hovering a new recent query', () => {
+    test('it should execute the search reflected in the search box', () => {});
+  });
+});